requirements.txt
asammdf
ipykernel
ipython
json5
jupyter-client
llama_index
matplotlib
mpld3
numpy
openai
pillow
streamlit_authenticator
streamlit==1.31.0 # So that the code intrepreter works!
lagent==0.1.2
